# Distributed Systems Programming/LAB01 - JSON design and REST API design/Server/api/openapi.yaml
openapi: 3.0.1
info:
  title: Film Library
  description: "A **simple application** to manage different films, requesting and\
    \ providing reviews and logging into a database full of contents! **Check it out!**"
  version: "1.0"
servers:
- url: http://localhost/
...
paths:
...
  /api/films/{filmId}:
    get:
      tags:
      - Film
      summary: Retrieve a single public or private film element.
      description: "A single public film can be viewed both by an authenticated user\
        \ and a non-authenticated one. On the other hand, if a film is marked as private,\
        \ a JWT identifying the owner must be provided. The film is identified by\
        \ its unique `filmId`."
      operationId: getFilm
      parameters:
      - name: filmId
        in: path
        description: The identifier of the film to be retrieved.
        required: true
        style: simple
        explode: false
        schema:
          type: integer
        example: 1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/film'
              example:
                filmId: 1
                title: Your Name
                owner: 1
                private: true
                watchdate: 2021-10-03
                rating: 9
                favorite: true
        "400":
          description: Bad Request
          content:
            text/plain:
              schema:
                type: string
                x-content-type: text/plain
              example: Invalid requestBody.
        "403":
          description: Forbidden
          content:
            text/plain:
              schema:
                type: string
                x-content-type: text/plain
              example: Must log in first.
        "404":
          description: Not Found
          content:
            text/plain:
              schema:
                type: string
                x-content-type: text/plain
              example: Film not found.
        "500":
          description: Server Error
          content:
            text/plain:
              schema:
                type: string
                x-content-type: text/plain
              example: Reason(s) for the server error.
        "505":
          description: Network Error
          content:
            text/plain:
              schema:
                type: string
                x-content-type: text/plain
              example: Reason(s) for the network error.
      x-swagger-router-controller: Film
    put:
      tags:
      - Film
      summary: Update an existing film.
      description: "An authenticated user can update an existing film, if she is the\
        \ owner of the film. However, this operation does not allow changing its visibility\
        \ from `public` to `private` (and viceversa)."
      operationId: editFilm
      parameters:
      - name: filmId
        in: path
        description: The identifier of the film to be modified.
        required: true
        style: simple
        explode: false
        schema:
          type: integer
        example: 2
      requestBody:
        description: The changes to be applied to the film.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/film'
            example:
              title: Godfather II
        required: true
      responses:
        "204":
          description: No content
          content: {}
        "400":
          description: Bad Request
          content:
            text/plain:
              schema:
                type: string
                x-content-type: text/plain
              example: Invalid requestBody.
        "401":
          description: Unauthorized
          content:
            text/plain:
              schema:
                type: string
                x-content-type: text/plain
              example: Must log in first.
        "403":
          description: Forbidden
          content:
            text/plain:
              schema:
                type: string
                x-content-type: text/plain
              example: Must log in first.
        "404":
          description: Not Found
          content:
            text/plain:
              schema:
                type: string
                x-content-type: text/plain
              example: Film not found.
        "500":
          description: Server error
          content:
            text/plain:
              schema:
                type: string
                x-content-type: text/plain
              example: Reason(s) for the server error.
        "505":
          description: Network Error
          content:
            text/plain:
              schema:
                type: string
                x-content-type: text/plain
              example: Reason(s) for the network error.
      security:
      - bearerAuth: []
      x-swagger-router-controller: Film
    delete:
      tags:
      - Film
      summary: Remove an existing film.
      description: "An authenticated user can delete an existing film, if she is the\
        \ owner of the film."
      operationId: removeFilm
      parameters:
      - name: filmId
        in: path
        description: The identifier of the film to be removed.
        required: true
        style: simple
        explode: false
        schema:
          type: string
        example: 1
      responses:
        "204":
          description: No content
          content: {}
        "400":
          description: Bad Request
          content:
            text/plain:
              schema:
                type: string
                x-content-type: text/plain
              example: Invalid requestBody.
        "401":
          description: Unauthorized
          content:
            text/plain:
              schema:
                type: string
                x-content-type: text/plain
              example: Must log in first.
        "500":
          description: Server error
          content:
            text/plain:
              schema:
                type: string
                x-content-type: text/plain
              example: Reason(s) for the server error.
        "505":
          description: Network Error
          content:
            text/plain:
              schema:
                type: string
                x-content-type: text/plain
              example: Reason(s) for the network error.
      security:
      - bearerAuth: []
      x-swagger-router-controller: Film
...
components:
  schemas:
...
    film:
      required:
      - filmId
      - owner
      - private
      - title
      type: object
      properties:
        filmId:
          maxLength: 36
          minLength: 36
          pattern: "[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}"
          type: string
          description: An unique identifier of the film data structure in the Film
            Manager service (mandatory).
        title:
          type: string
          description: Textual title of the film (mandatory).
        owner:
          pattern: "USER_[0-9]{10}"
          type: string
          description: "The id of the film data structure owner, i.e., the user who\
            \ created it (mandatory)."
        private:
          type: boolean
          description: "A mandatory Boolean property, set to true if the film data\
            \ structure is marked as private, false if it is public (default value:\
            \ true). A film data structure is said private if only its owner can access\
            \ it, public if every user can access it"
        watchdate:
          type: string
          description: "The date when the film was watched by the owner, expressed\
            \ in the YYYY-MM-DD format."
          format: date
        rating:
          maximum: 10
          minimum: 0
          type: integer
          description: This property can be included in the film data structure only
            if private is true.
        favorite:
          type: boolean
          description: This property can be included in the film data structure only
            if private is true.
          default: false
      example:
        owner: owner
        watchdate: 2000-01-23
        private: true
        filmId: filmId
        rating: 5
        title: title
        favorite: false
...
  securitySchemes:
    bearerAuth:
      type: http
      description: JWT to identify the authenticated user.
      scheme: bearer
      bearerFormat: JWT